rename lines to arcs
